Implemented BigDecimal in calculator

+++ Calculator/src/main/java/nl/ordina/academy/calculator/Calculator.java
@@ -2,6 +2,8 @@ package nl.ordina.academy.calculator;
 
 import nl.ordina.academy.calculator.exception.*;
 
+import java.math.BigDecimal;
+
 /**
  * @author Dirk Luijk <dirk.luijk@ordina.nl>
  */
@@ -21,7 +23,7 @@ public class Calculator {
         this(new Tokenizer(), new RPNConverter(), new RPNInterpreter());
     }
 
-    public double calculate(String input) throws CalculatorException {
+    public BigDecimal calculate(String input) throws CalculatorException {
 
         if (input.isEmpty()) {
             throw new EmptyInputException("Input may not be empty.");
